system-image-upgrader: Fix erroneous vendor symlink
Turns out that, currently, system-images contain twice the amount of
symlinks to their /android counterparts, which breaks expectations of
the system-image-upgrader being able to install a working system.

It's both rootfs-builder-debos and system-image-server creating symlinks,
a simple 'tar -tf ubports-tarball.tar.xz | grep vendor' shows:
system/./vendor
system/vendor

So, as a quick workaround, remove and replace the final wrong symlink
with the correct one. Let's revert this change as soon as this situation
is sorted out and the tarball only includes the valid vendor symlink.

Change-Id: Idf1fea7568c8a59e8d7d96349014e14d6db51a8f