(ubports) init: set property to identify ubports recovery
The UBports installer has no way of knowing, if the recovery currently
used is a UBports specific recovery, which is required in order to
install Ubuntu Touch via system-image-upgrader.

Set a property which we can check for with ADB to detect if a recovery
supporting system-image-upgrader is running.

Issue: https://github.com/ubports/halium_bootable_recovery/issues/22

Change-Id: I0f2a00a0a8f9b68203804f1c61f9b639cf7227fc
Signed-off-by: Alexander Martinz <amartinz@shiftphones.com>